Fix deadlock and signed integer overflow found by Devin Review
- DiagnosticSystem::generate_diagnostic_report(): inline recent events
  logic instead of calling get_recent_events() which would deadlock on
  the already-held non-recursive diagnostic_mutex
- MemoryRegion::read_dword(): cast uint8_t to uint32_t before shifting
  to avoid signed integer overflow (undefined behavior) when high byte
  >= 0x80

Refs: OZC-274

+++ libgnucash/engine/gnc-virtual-device.cpp
@@ -23,8 +23,8 @@ uint16_t MemoryRegion::read_word(uint64_t offset) {
 
 uint32_t MemoryRegion::read_dword(uint64_t offset) {
     if (offset + 3 >= size) return 0;
-    return (data[offset + 3] << 24) | (data[offset + 2] << 16) | 
-           (data[offset + 1] << 8) | data[offset];
+    return (static_cast<uint32_t>(data[offset + 3]) << 24) | (static_cast<uint32_t>(data[offset + 2]) << 16) | 
+           (static_cast<uint32_t>(data[offset + 1]) << 8) | data[offset];
 }
 
 void MemoryRegion::write_byte(uint64_t offset, uint8_t value) {
